// src/generated/client/models/TruckDriveState.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { TruckDriveStateEnum } from './TruckDriveStateEnum';
import {
    TruckDriveStateEnumFromJSON,
    TruckDriveStateEnumFromJSONTyped,
    TruckDriveStateEnumToJSON,
} from './TruckDriveStateEnum';

/**
 * 
 * @export
 * @interface TruckDriveState
 */
export interface TruckDriveState {
    /**
     * Truck drive state ID
     * @type {string}
     * @memberof TruckDriveState
     */
    readonly id?: string;
    /**
     * Timestamp for truck drive state
     * @type {number}
     * @memberof TruckDriveState
     */
    timestamp: number;
    /**
     * 
     * @type {TruckDriveStateEnum}
     * @memberof TruckDriveState
     */
    state: TruckDriveStateEnum;
    /**
     * Driver ID.
     * @type {string}
     * @memberof TruckDriveState
     */
    readonly driverId?: string;
    /**
     * Driver card ID. This is used to identify the driver when creating a new drive state.
     * @type {string}
     * @memberof TruckDriveState
     */
    driverCardId?: string;
}

/**
 * Check if a given object implements the TruckDriveState interface.
 */
export function instanceOfTruckDriveState(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "timestamp" in value;
    isInstance = isInstance && "state" in value;

    return isInstance;
}

export function TruckDriveStateFromJSON(json: any): TruckDriveState {
    return TruckDriveStateFromJSONTyped(json, false);
}

export function TruckDriveStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TruckDriveState {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'timestamp': json['timestamp'],
        'state': TruckDriveStateEnumFromJSON(json['state']),
        'driverId': !exists(json, 'driverId') ? undefined : json['driverId'],
        'driverCardId': !exists(json, 'driverCardId') ? undefined : json['driverCardId'],
    };
}

export function TruckDriveStateToJSON(value?: TruckDriveState | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'timestamp': value.timestamp,
        'state': TruckDriveStateEnumToJSON(value.state),
        'driverCardId': value.driverCardId,
    };
}